pineapple: Add a config property for the SMMU proxy driver
Add a board config property to indicate that the SMMU proxy driver is
present on this vendor SI in the securemsm-kernel folder. The display
SI depends on symbols from the SMMU proxy driver, such that the
display SI is also used across multiple targets, some of which will
not contain the SMMU proxy driver. Expose TARGET_USES_SMMU_PROXY to
allow the display SI to control at compile time whether they should
use symbols from the proxy driver or not.

Change-Id: Ibb94849d7fa2d9078c95a03ce2561973556c6119