device: pineapple: Increase DTBO partition size to 12 MB
With the increase in new device tree overlay files
introduced by Cliffs SoC, which is binary compatible
to Pineapple SoC, the size of DTBO partition reaches
its maximum size, which leads to compilation failures.
Increase DTBO partition size for Pineapple from 8 MB
to 12 MB to handle this and accomodate the new overlay
files that are going to be introduced in future.

Change-Id: I7461b6b9cbd3aa755726d128fc6ef78a1a0264f0