device: pineapple: Increase DTBO partition size to 16 MB
With the increase in new device tree overlay files
introduced by Cliffs, Volcano SoC, which are binary
compatible to Pineapple SoC, the size of DTBO partition
reaches its maximum size, which leads to compilation
failures.
Increase DTBO partition size for Pineapple from 12 MB
to 16 MB to handle this and accomodate the new overlay
files that are going to be introduced in future.

Change-Id: Idfe63a35cac2c6ae096714ff4ec4dfd721813a68