device: pineapple: Increase DTBO partition size to 20 MB
With the increase in new device tree overlay files
introduced by Cliffs, Volcano SoC, which are binary
compatible to Pineapple SoC, the size of DTBO partition
reaches its maximum size, which leads to compilation
failures.
Increase DTBO partition size for Pineapple from 16 MB
to 20 MB to handle this and accomodate the new overlay
files that are going to be introduced in future.

Change-Id: I8d4c828f124c7c20607c06592d8f2e8c123503ad